refactor(cli): rename pkg/upgrade to pkg/selfupdate (DEVSY-016)
